requirements.txt
aioredis==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
autobahn==23.1.2
Automat==22.10.0
backports.zoneinfo==0.2.1
certifi==2022.12.7
cffi==1.15.1
channels==4.0.0
channels-redis==4.0.0
charset-normalizer==3.0.1
chess==1.9.4
constantly==15.1.0
cryptography==40.0.1
daphne==2.5.0
Django==4.1.5
django-ckeditor==6.5.1
django-cors-headers==3.13.0
django-js-asset==2.0.0
djangorestframework==3.14.0
hiredis==2.2.2
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
Jinja2==3.1.2
MarkupSafe==2.1.2
msgpack==1.0.5
numpy==1.24.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==23.1.0
pytz==2022.7.1
redis==4.4.2
requests==2.28.2
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.3
stockfish==3.28.0
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.5.0
urllib3==1.26.14
zope.interface==6.0
